Refactor computation + problem def
